build(nx): migrate codegen:readme to @clc/nx
