+++ packages/@clc/nx/executors/readme/schema.d.ts
@@ -1,0 +1,8 @@
+/* eslint-disable */
+/**
+ * This file was automatically generated by json-schema-to-typescript.
+ * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+ * and run json-schema-to-typescript to regenerate this file.
+ */
+
+export interface ReadmeExecutor {}
