+++ packages/@clc/nx/executors/readme/schema.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "version": 2,
+  "title": "Readme executor",
+  "description": "",
+  "type": "object",
+  "properties": {},
+  "required": [],
+  "additionalProperties": false
+}
