+++ packages/@code-like-a-carpenter/nx-auto/package.json
@@ -2,13 +2,7 @@
   "name": "@code-like-a-carpenter/nx-auto",
   "main": "./src/index.js",
   "dependencies": {
-    "mdast": "^3.0.0",
     "prettier": "^3.2.5",
-    "remark": "^14.0.2",
-    "remark-stringify": "^10.0.2",
-    "remark-toc": "^8.0.1",
-    "unified": "^10.1.2",
-    "unist-builder": "^3.0.1",
     "yargs": "^17.7.2"
   },
   "devDependencies": {
