--- packages/@code-like-a-carpenter/nx-auto/src/readme/index.js
@@ -1,4 +1,0 @@
-'use strict';
-
-exports.ensureHeader = require('./ensure-header').ensureHeader;
-exports.ensureSections = require('./ensure-sections').ensureSections;
